Move Request to Http namespace
